+++ src/IdBrokerClient.php
@@ -23,11 +23,11 @@ class IdBrokerClient extends BaseClient
      *
      * @var RangeInterface[]
      */
-    private $trustedIpRanges = [ ];
+    private array $trustedIpRanges = [];
 
-    private $assertValidBrokerIp = true;
+    private bool $assertValidBrokerIp = true;
 
-    private $idBrokerUri;
+    private string $idBrokerUri;
 
     /**
      * Constructor.
@@ -88,7 +88,7 @@ class IdBrokerClient extends BaseClient
      * @throws \InvalidArgumentException
      * @throws \Exception if assertValidBrokerIp is true and idBrokerUri is invalid, unresolvable, or untrusted
      */
-    private function initializeConfig($config)
+    private function initializeConfig(array $config)
     {
 
         if (isset($config[ self::ASSERT_VALID_BROKER_IP_CONFIG ])) {
@@ -144,7 +144,7 @@ class IdBrokerClient extends BaseClient
      * @return array|null An array of user information (if valid), or null.
      * @throws ServiceException
      */
-    public function authenticate(string $username, string $password)
+    public function authenticate(string $username, string $password): ?array
     {
         $result = $this->authenticateInternal([
             'username' => $username,
@@ -168,7 +168,7 @@ class IdBrokerClient extends BaseClient
      * @return array|null An array of user information (if valid), or null.
      * @throws ServiceException
      */
-    public function authenticateNewUser(string $invite)
+    public function authenticateNewUser(string $invite): ?array
     {
         $result = $this->authenticateNewUserInternal([
             'invite' => $invite,
@@ -192,7 +192,7 @@ class IdBrokerClient extends BaseClient
      * @return array An array of information about the new user.
      * @throws ServiceException
      */
-    public function createUser(array $config = [ ])
+    public function createUser(array $config = [ ]): array
     {
         $result = $this->createUserInternal($config);
         $statusCode = (int)$result[ 'statusCode' ];
@@ -229,7 +229,7 @@ class IdBrokerClient extends BaseClient
      * @param Result $result The result of a Guzzle call.
      * @return array
      */
-    protected function getResultAsArrayWithoutStatusCode($result)
+    protected function getResultAsArrayWithoutStatusCode(Result $result): array
     {
         unset($result[ 'statusCode' ]);
         return $result->toArray();
@@ -241,7 +241,7 @@ class IdBrokerClient extends BaseClient
      * @return string "OK".
      * @throws ServiceException
      */
-    public function getSiteStatus()
+    public function getSiteStatus(): string
     {
         $result = $this->getSiteStatusInternal();
         $statusCode = (int)$result[ 'statusCode' ];
@@ -261,7 +261,7 @@ class IdBrokerClient extends BaseClient
      *     null if no such user was found.
      * @throws ServiceException
      */
-    public function getUser(string $employeeId)
+    public function getUser(string $employeeId): ?array
     {
         $result = $this->getUserInternal([
             'employee_id' => $employeeId,
@@ -282,12 +282,12 @@ class IdBrokerClient extends BaseClient
      *
      * @param array|null $fields (Optional:) The list of fields desired about
      *     each user in the result.
-     * @param  array|null $search (Optional:) An array of fields to search on,
+     * @param array|null $search (Optional:) An array of fields to search on,
      *     example ['username' => 'billy']
      * @return array An array with a sub-array about each user.
      * @throws ServiceException
      */
-    public function listUsers($fields = null, $search = [ ])
+    public function listUsers(array $fields = null, ?array $search = []): array
     {
         $config = [ ];
         if ($fields !== null) {
@@ -308,16 +308,18 @@ class IdBrokerClient extends BaseClient
      * Create a new MFA configuration
      * @param string $employee_id
      * @param string $type
-     * @param string $label
+     * @param string|null $label
+     * @param string|null $rpOrigin
      * @return array|null
      * @throws ServiceException
      */
-    public function mfaCreate($employee_id, $type, $label = null)
+    public function mfaCreate(string $employee_id, string $type, string $label = null, string $rpOrigin = ''): ?array
     {
         $result = $this->mfaCreateInternal([
             'employee_id' => $employee_id,
             'type' => $type,
             'label' => $label,
+            'rpOrigin' => $rpOrigin,
         ]);
         $statusCode = (int)$result[ 'statusCode' ];
 
@@ -335,7 +337,7 @@ class IdBrokerClient extends BaseClient
      * @return null
      * @throws ServiceException
      */
-    public function mfaDelete($id, $employeeId)
+    public function mfaDelete(int $id, string $employeeId)
     {
         $result = $this->mfaDeleteInternal([
             'id' => $id,
@@ -353,13 +355,15 @@ class IdBrokerClient extends BaseClient
     /**
      * Get a list of MFA configurations for given user
      * @param string $employee_id
+     * @param string $rpOrigin
      * @return array
      * @throws ServiceException
      */
-    public function mfaList($employee_id)
+    public function mfaList(string $employee_id, string $rpOrigin = ''): array
     {
         $result = $this->mfaListInternal([
             'employee_id' => $employee_id,
+            'rpOrigin' => $rpOrigin,
         ]);
         $statusCode = (int)$result[ 'statusCode' ];
 
@@ -378,7 +382,7 @@ class IdBrokerClient extends BaseClient
      * @return array
      * @throws ServiceException
      */
-    public function mfaUpdate($id, $employeeId, $label)
+    public function mfaUpdate(int $id, string $employeeId, string $label): array
     {
         $result = $this->mfaUpdateInternal([
             'id' => $id,
@@ -403,12 +407,13 @@ class IdBrokerClient extends BaseClient
      * @throws MfaRateLimitException
      * @throws ServiceException
      */
-    public function mfaVerify($id, $employeeId, $value)
+    public function mfaVerify(string $id, string $employeeId, string $value, string $rpOrigin = '')
     {
         $result = $this->mfaVerifyInternal([
             'id' => $id,
             'employee_id' => $employeeId,
             'value' => $value,
+            'rpOrigin' => $rpOrigin,
         ]);
         $statusCode = (int)$result[ 'statusCode' ];
 
@@ -434,7 +439,7 @@ class IdBrokerClient extends BaseClient
      * @return String[]
      * @throws ServiceException
      */
-    public function createMethod($employee_id, $value, $created = '')
+    public function createMethod(string $employee_id, string $value, string $created = ''): array
     {
         $params = compact('employee_id', 'value');
         if (! empty($created)) {
@@ -458,7 +463,7 @@ class IdBrokerClient extends BaseClient
      * @return null
      * @throws ServiceException
      */
-    public function deleteMethod($uid, $employee_id)
+    public function deleteMethod(int $uid, int $employee_id)
     {
         $result = $this->deleteMethodInternal(compact('uid', 'employee_id'));
         $statusCode = (int)$result[ 'statusCode' ];
@@ -477,7 +482,7 @@ class IdBrokerClient extends BaseClient
      * @return String[]
      * @throws ServiceException
      */
-    public function getMethod($uid, $employee_id)
+    public function getMethod(int $uid, int $employee_id): array
     {
         $result = $this->getMethodInternal(compact('uid', 'employee_id'));
         $statusCode = (int)$result[ 'statusCode' ];
@@ -495,7 +500,7 @@ class IdBrokerClient extends BaseClient
      * @return String[]
      * @throws ServiceException
      */
-    public function listMethod($employee_id)
+    public function listMethod(string $employee_id): array
     {
         $result = $this->listMethodInternal(compact('employee_id'));
         $statusCode = (int)$result[ 'statusCode' ];
@@ -515,7 +520,7 @@ class IdBrokerClient extends BaseClient
      * @return String[]
      * @throws ServiceException
      */
-    public function verifyMethod($uid, $employee_id, $code)
+    public function verifyMethod(string $uid, string $employee_id, string $code): array
     {
         $result = $this->verifyMethodInternal(compact('uid', 'employee_id', 'code'));
         $statusCode = (int)$result[ 'statusCode' ];
@@ -534,7 +539,7 @@ class IdBrokerClient extends BaseClient
      * @return bool
      * @throws ServiceException
      */
-    public function resendMethod($uid, $employee_id)
+    public function resendMethod(string $uid, string $employee_id): bool
     {
         $result = $this->resendMethodInternal(compact('uid', 'employee_id'));
         $statusCode = (int)$result[ 'statusCode' ];
@@ -556,7 +561,7 @@ class IdBrokerClient extends BaseClient
      * @return array An array of password metadata
      * @throws ServiceException
      */
-    public function setPassword(string $employeeId, string $password)
+    public function setPassword(string $employeeId, string $password): array
     {
         $result = $this->setPasswordInternal([
             'employee_id' => $employeeId,
@@ -581,7 +586,7 @@ class IdBrokerClient extends BaseClient
      * @return bool
      * @throws ServiceException
      */
-    public function assessPassword(string $employeeId, string $password)
+    public function assessPassword(string $employeeId, string $password): bool
     {
         $result = $this->assessPasswordInternal([
             'employee_id' => $employeeId,
@@ -597,11 +602,11 @@ class IdBrokerClient extends BaseClient
     }
 
     /**
-     * @param \GuzzleHttp\Command\Result $response
+     * @param Result $response
      * @param int $uniqueErrorCode
      * @throws ServiceException
      */
-    protected function reportUnexpectedResponse($response, $uniqueErrorCode)
+    protected function reportUnexpectedResponse(Result $response, int $uniqueErrorCode)
     {
         throw new ServiceException(
             sprintf(
@@ -621,7 +626,7 @@ class IdBrokerClient extends BaseClient
      * @return array An array of information about the updated user.
      * @throws ServiceException
      */
-    public function updateUser(array $config = [ ])
+    public function updateUser(array $config = [ ]): array
     {
         $result = $this->updateUserInternal($config);
         $statusCode = (int)$result[ 'statusCode' ];
@@ -660,7 +665,7 @@ class IdBrokerClient extends BaseClient
      * @param string $ipAddress The IP address in question.
      * @return bool
      */
-    private function isTrustedIpAddress($ipAddress)
+    private function isTrustedIpAddress(string $ipAddress): bool
     {
         foreach ($this->trustedIpRanges as $trustedIpBlock) {
             $addr = Factory::addressFromString($ipAddress);