+++ src/ServiceException.php
@@ -8,9 +8,15 @@ class ServiceException extends \Exception
     /**
      * @var null|int The status code returned by ID Broker API
      */
-    public $httpStatusCode;
+    public ?int $httpStatusCode;
 
-    public function __construct($message = "", $code = 0, $httpStatusCode = null, Throwable $previous = null)
+    /**
+     * @param string $message
+     * @param int $code
+     * @param int|null $httpStatusCode
+     * @param Throwable|null $previous
+     */
+    public function __construct(string $message = "", int $code = 0, ?int $httpStatusCode = null, Throwable $previous = null)
     {
         $this->httpStatusCode = $httpStatusCode;
         parent::__construct($message, $code, $previous);